pyproject.toml
[tool.poetry]
name = "Talouden seuranta"
version = "0.1.0"
description = ""
authors = ["Tiina <tiina.kuisma@helsinki.fi>"]

[tool.poetry.dependencies]
python = "^3.8"
invoke = "^1.6.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
coverage = "^6.1.2"
pylint = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
